feat: add 3 classes for use case
